[Event] Move Provider class to abstract namespace.
