+++ src/Valkyrja/Event/Provider/Abstract/Provider.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Event\Provider;
+namespace Valkyrja\Event\Provider\Abstract;
 
 use Override;
 use Valkyrja\Event\Provider\Contract\ProviderContract;
